% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtgis.R
\name{mtgis_counties}
\alias{mtgis_counties}
\title{Get Montana county boundaries (attributes only)}
\usage{
mtgis_counties(max_records = 100)
}
\arguments{
\item{max_records}{Maximum records (default 100)}
}
\value{
tibble of county attributes: OBJECTID, NAME, FIPS, Acres, SQMILES, etc.
}
\description{
Get Montana county boundaries (attributes only)
}
