% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtgis.R
\name{mtgis_layers}
\alias{mtgis_layers}
\title{List layers in a Montana GIS map service}
\usage{
mtgis_layers(service)
}
\arguments{
\item{service}{Service path (e.g. "MSDI_Framework/Boundaries")}
}
\value{
tibble: id, name, type, geometryType, minScale, maxScale
}
\description{
List layers in a Montana GIS map service
}
