% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtgis.R
\name{mtgis_services}
\alias{mtgis_services}
\title{List available Montana GIS map services}
\usage{
mtgis_services(folder = NULL)
}
\arguments{
\item{folder}{Optional folder name (e.g. "MSDI_Framework", "DLI", "MSL")}
}
\value{
tibble: name, type
}
\description{
List available Montana GIS map services
}
